Testing

+++ Form1.cs
@@ -14,39 +14,60 @@ namespace HPCProject
     public partial class Form1 : Form
     {
         public string searchQuery;
-        public int searchMode;
-        public int threadsNumber;
-        public List<int> results = new List<int>();
+        public mode searchMode;
+        public int num_threads;
+        int length;
+        int[] positions;
+
         public Form1()
         {
             InitializeComponent();
-            //////////////testing 
-            int x = Wrap.subtract(10, 5);
-            string tmp = "islam omar islam islam";
-            IntPtr textsearch = Wrap.Create_TextSearch_Obj(tmp); // 
-            int length;
-            int[] positions;
-            Wrap.Execute_SearchforWord(textsearch, "islam", mode.Serial, out positions, out length);
-            double time = Wrap.Execute_GetElapsedTime(textsearch);
-            Wrap.Delete_TextSearch_Obj(textsearch); //don't forget
-            int vx = 0;
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        private void searchButton_Click(object sender, EventArgs e)
         {
+            if (!IsDataComplete())
+            {
+                MessageBox.Show("Please Enter Search Query.");
+                return;
+            }
+
+
             searchQuery = this.searchQueryTextBox.Text;
-            threadsNumber = Convert.ToInt32(this.numOfThreadsNumericUpDown.Value);
-            if(this.parallelRadioButton.Checked==true)
-            {
-                this.searchMode = 0;
-            }
-            else
-            {
-                this.searchMode = 1;
-            }
+            num_threads = Convert.ToInt32(this.numOfThreadsNumericUpDown.Value);
+            
+            IntPtr textsearch = Wrap.Create_TextSearch_Obj(searchQuery);
+            Wrap.Execute_SearchforWord(textsearch, searchQuery, searchMode, out positions, out length);
+            double time = Wrap.Execute_GetElapsedTime(textsearch);
+            Wrap.Delete_TextSearch_Obj(textsearch); //don't forget
+        }
+
+        private void DisplayResults()
+        {
+            resultsListBox.Items.Add(positions);
+        }
+
+        private bool IsDataComplete()
+        {
+            return searchQueryTextBox.Text.Length != 0;
+        }
+
+        private void parallelRadioButton_CheckedChanged(object sender, EventArgs e)
+        {
+            searchMode = mode.Parallel;
+        }
+
+        private void sequentialRadioButton_CheckedChanged(object sender, EventArgs e)
+        {
+            searchMode = mode.Serial;
+        }
+
+        private void parallelGPURadioButton_CheckedChanged(object sender, EventArgs e)
+        {
+            searchMode = mode.ParallelGPU;
         }
     }
-    public enum mode { Parallel, Serial };
+    public enum mode { Parallel, Serial, ParallelGPU };
     public class Wrap
     {
         [DllImport(@"D:\Text-Search-master\Release\TextSearchCpluse.dll", CallingConvention = CallingConvention.Cdecl)]
@@ -59,7 +80,5 @@ namespace HPCProject
         public static extern double Execute_GetElapsedTime(IntPtr Obj);
         [DllImport(@"D:\Text-Search-master\Release\TextSearchCpluse.dll", CallingConvention = CallingConvention.Cdecl)]
         public static extern void Execute_SearchforWord(IntPtr Obj, string keyword, mode searchMode, [MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 4)] out int[] data, out int count);
-
     }
-
 }
